refactor(post-in-memory.repository.ts): rename
